backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "joblib>=1.5.1",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.0.0",
    "sqlalchemy>=2.0.41",
    "torch>=2.7.1",
    "transformers>=4.53.2",
    "uvicorn>=0.35.0",
    "xgboost>=3.0.2",
]
